Rename filename related to firebase
