+++ src/containers/user/components/Guest.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, StyleSheet, } from 'react-native';
 import colors from '../../../theme/colors';
-import { googleLoginAndLink } from '../../../firebase/Auth';
+import { googleLoginAndLink } from '../../../service/AuthService';
 
 const GuestInfo = () => {
     return (
