--- src/firebase/Handle.js
@@ -1,10 +1,0 @@
-import uploadImageToStorage from "./Storage";
-import { createGroup, createTiccle } from "./Firestore";
-
-function handleCreateGroup() {
-    
-}
-
-function handleCreateTiccle() {
-
-}
