+++ src/service/GroupService.js
@@ -1,6 +1,6 @@
 import firestore from '@react-native-firebase/firestore';
-import { getCurrentUser } from './Auth';
-import { uploadImageToStorage, getDownloadURLByName } from './Storage';
+import { getCurrentUser } from './AuthService';
+import { uploadImageToStorage, getDownloadURLByName } from './ImageService';
 
 const user = getCurrentUser();
 const userDoc = firestore().collection('RTiccle').doc(user.uid);
@@ -16,16 +16,6 @@ async function createGroup(groupName, newGroup) {
     console.log(newGroup);
     await ref.set(newGroup);
     return ref.id;
-}
-
-/**
- * @param {*} newTiccle
- * @returns {string} Ticcle Id
- */
-async function createTiccle(newTiccle) {
-    const ref = userDoc.collection("Ticcle"); // using auto id
-    const ticcleRef = await ref.add(newTiccle);
-    return ticcleRef.id;
 }
 
 /**
@@ -49,35 +39,6 @@ function uploadNewGroup(newGroupName, group, mainImageSource) {
         uploadImageToStorage(imageName, mainImageSource);
     }
     return createGroup(newGroupName, { ...group, mainImage: imageName });
-}
-
-/**
- * Upload new ticcle to firestore (upload image and group)
- * @param {*} ticcle: ticcle info 
- *  * {
-        lastModifiedTime: TimeStamp,
-        group: gid,
-        title: String,
-        link: String, // URL of original content
-        content: String,
-        tagList: Array<String>
-    }
- * @param {Array} images: image source array - LIMIT 2
- * @returns {Promise<String>} created ticcle id (random)
- */
-function uploadNewTiccle(ticcle, images) {
-    // upload images first
-    var imageArr = [];
-    if (images !== undefined) {
-        images.map((image) => {
-            const imageName = Date.now() + ".jpg";
-            imageArr.push(imageName);
-            uploadImageToStorage(imageName, image);
-        })
-    }
-
-    // upload ticcle info
-    return createTiccle({ ...ticcle, images: imageArr });
 }
 
 /**
@@ -145,52 +106,6 @@ async function findGroupById(groupId) {
 }
 
 /**
- * Get ticcle list by group id
- * @param {string} groupId 
- * @returns {Array} Ticcle List
- */
-async function findTiccleListByGroupId(groupId) {
-    const query = userDoc.collection("Ticcle")
-        .where("group", "==", groupId);
-    const querySnapshot = await query.get();
-
-    var ticcleList = [];
-    querySnapshot.forEach(snapshot => {
-        const id = snapshot.id;
-        const ticcle = { ...snapshot.data(), id }
-        ticcleList = [...ticcleList, ticcle];
-    });
-    return ticcleList;
-}
-
-/**
- * Get One Ticcle By Id (DocumentSnapshot.id)
- * @param {*} ticcleId 
- * @returns {DocumentSnapshot} (of Ticcle doc) if exist, else null
- */
-async function findTiccleById(ticcleId) {
-    const ticcle = await userDoc.collection("Ticcle").doc(ticcleId).get()
-    if (ticcle.exists) return ticcle.data();
-    else return null;
-}
-
-/**
- * Get images of ticcle
- * @param {Array} images // image name array - LIMIT 2
- * @returns {Array} Image URL Array
- */
-async function getImagesOfTiccle(images) {
-    var imageURLArr = [];
-    if (images !== undefined) {
-        images.map((imageName) => {
-            const URL = getDownloadURLByName(imageName, true);
-            imageURLArr.push(URL);
-        })
-    }
-    return imageURLArr;
-}
-
-/**
  * Check existed group
  * @returns {boolean} true: existed, false: not existed
  */
@@ -201,15 +116,6 @@ async function checkExsitedGroup() {
     } else {
         return true;
     }
-}
-
-/**
- * get group's ticcle count by groupId
- * @returns {Int} ticcle length
- */
-async function getTiccleCount(groupId) {
-    const ticcleList = await findTiccleListByGroupId(groupId);
-    return ticcleList.length;
 }
 
 /**
@@ -229,17 +135,11 @@ async function getGroupDataIncludeImage(groupId) {
 
 export {
     createGroup,
-    createTiccle,
     uploadNewGroup,
-    uploadNewTiccle,
     findAllGroup,
     findGroupsIncludeImage,
     checkIsExistingGroup,
-    findTiccleListByGroupId,
     findGroupById,
-    findTiccleById,
-    getImagesOfTiccle,
     checkExsitedGroup,
-    getTiccleCount,
     getGroupDataIncludeImage,
 }
